refactor: move shared helpers to app/ and colocate single-file components into their own folders
